# feat(http): gzip compress the query CSV response.